Changes allowing to get rid of virtual
This commit allows working with AOTRelocationsWIP, by making some of
methods within WASMRelocationRecords virtual, which should not be like
that.
PreparePrivateData was a problem to be not virtual, while
applyRelocations needed to be virtual. Both were replaced with a dispat-
cher, which looks a bit ugly, but seems to work.
Ongoing work on fixing the size of AOTRelocationRecords sizes and
application.
This commit also contains changes to relocations, removing a nonexistent
kind.

Signed-off-by: Georgiy Krylov <georgiy.krylov@unb.ca>