Changes to facilitate counting imports
The import functions are now counted as static variable. Important to
note that the code added by emscripten (em-interp) is added through add
host export, hence this import counting does not work for that. The
workaround was to count them in the compiler library. Another issue that
might appear is the code might fail to run properly as there might be
issues with multiple modules requiring imports (static variable is
likely not the best way to count import across all the modules).