Changes to import counting
Previously in wabtaot the imports were triggering
BinaryReaderInterp::OnImportFunc, whereas em-interp defines them using
AppendFuncExport, which triggers onHostFunctionExport?? callback. This
callback, however, does not trigger an import. The AOT compiler now
features WebAssembly-specific function to count the imports and distin-
guishing between modules using the exports count and function offset.
There was an issue that functions that are not exported do not
 contribute towards the number of exports, as well as the number of im-
 ports if multiple modules are loaded is considered the same (which is
 a fair assumption, as emscripten defines global imports like memories
 and some OS functions).

Signed-off-by: Georgiy Krylov <georgiy.krylov@unb.ca>