Changes to accommodate AOTRelocationsWIP changes
The Relocation Private Data is supposed to be a union of PrivateData
types. In the previous version, it was just an integer.
AOTMethodHeader was also changed, hence, updating the code using it.
RelocationRuntime related code was also updated to delete OMRAOTReloca-
tionsRuntime.
TR_RelocationTarget do not have record value anymore and uses pointers
instead.
TR_AOTStorageInterface is renamed into TR::AOTStorage.

Signed-off-by: Georgiy Krylov <georgiy.krylov@unb.ca>