Changes to improve call
Currently, the call works as it was designed: bottom up traversal when
compiling, thus leading to different methods being compiled at diff
time.
However, there's an issue of the module name being hard coded.
Another unaddressed issue is recursive methods calling each other,
with cyclic graphs of dependencies.
One more issue is to check if ELF is re-emitted when it's not necessary.

Signed-off-by: Georgiy Krylov <georgiy.krylov@unb.ca>